test/inspector/runtime/console-memory-expected.txt

Checks console.memory

Running test: testWithoutMemory
{
    id : <messageId>
    result : {
        result : {
            type : undefined
        }
    }
}

Running test: testSetterInStrictMode
{
    id : <messageId>
    result : {
        result : {
            type : undefined
        }
    }
}

Running test: testWithMemory
{
    id : <messageId>
    result : {
        result : {
            type : number
            value : 239
        }
    }
}

Running test: testSetterDoesntOverride
{
    id : <messageId>
    result : {
        result : {
            type : number
            value : 42
        }
    }
}
